+++ tests/components/supernotify/support_cases/test_issue_36_explicit_selection.py
@@ -56,8 +56,8 @@ async def test_mobile_push_only_when_no_explicit_delivery(support_case_fixture, 
     await uut.initialize()
     await uut.deliver()
 
-    assert list(uut.deliveries.keys()) == unordered("signal", "DEFAULT_mobile_push")
-    assert len(uut.deliveries["DEFAULT_mobile_push"][EnvelopeOutcome.SUCCESS]) == 1
+    assert list(uut.deliveries.keys()) == unordered("signal", "mobile_push")
+    assert len(uut.deliveries["mobile_push"][EnvelopeOutcome.SUCCESS]) == 1
     assert len(uut.deliveries["signal"][EnvelopeOutcome.SUCCESS]) == 1
 
 
@@ -77,9 +77,9 @@ async def test_explicit_delivery_and_mobile_and_implicit_selection(support_case_
     await uut.initialize()
     await uut.deliver()
 
-    assert list(uut.deliveries.keys()) == unordered("signal", "DEFAULT_mobile_push")
+    assert list(uut.deliveries.keys()) == unordered("signal", "mobile_push")
     assert len(uut.deliveries["signal"][EnvelopeOutcome.SUCCESS]) == 1
-    assert len(uut.deliveries["DEFAULT_mobile_push"][EnvelopeOutcome.SUCCESS]) == 1
+    assert len(uut.deliveries["mobile_push"][EnvelopeOutcome.SUCCESS]) == 1
 
 
 async def test_explicit_delivery_no_mobile(support_case_fixture, hass: HomeAssistant):
@@ -88,6 +88,6 @@ async def test_explicit_delivery_no_mobile(support_case_fixture, hass: HomeAssis
     await uut.initialize()
     await uut.deliver()
 
-    assert list(uut.deliveries.keys()) == unordered("signal", "DEFAULT_mobile_push")
+    assert list(uut.deliveries.keys()) == unordered("signal", "mobile_push")
     assert len(uut.deliveries["signal"][EnvelopeOutcome.SUCCESS]) == 1
-    assert len(uut.deliveries["DEFAULT_mobile_push"].get(EnvelopeOutcome.SUCCESS, [])) == 0
+    assert len(uut.deliveries["mobile_push"].get(EnvelopeOutcome.SUCCESS, [])) == 0