+++ tests/components/supernotify/test_notify_snooze.py
@@ -6,10 +6,10 @@ from homeassistant.core import Context, Event, HomeAssistant
 from custom_components.supernotify.const import (
     ATTR_ACTION,
     ATTR_USER_ID,
+    CONF_INCLUSION,
     CONF_PERSON,
-    CONF_SELECTION,
     CONF_TRANSPORT,
-    SELECTION_BY_SCENARIO,
+    INCLUSION_BY_SCENARIO,
     TRANSPORT_ALEXA_MEDIA_PLAYER,
     TRANSPORT_CHIME,
     TRANSPORT_EMAIL,
@@ -31,7 +31,7 @@ DELIVERY: dict[str, dict] = {
     "chime": {CONF_TRANSPORT: TRANSPORT_CHIME, "target": ["switch.bell_1", "script.siren_2"]},
     "alexa_media_player": {CONF_TRANSPORT: TRANSPORT_ALEXA_MEDIA_PLAYER, CONF_ACTION: "notify.alexa_media_player"},
     "chat": {CONF_TRANSPORT: TRANSPORT_GENERIC, CONF_ACTION: "notify.my_chat_server"},
-    "persistent": {CONF_TRANSPORT: TRANSPORT_PERSISTENT, CONF_SELECTION: [SELECTION_BY_SCENARIO]},
+    "persistent": {CONF_TRANSPORT: TRANSPORT_PERSISTENT, CONF_INCLUSION: [INCLUSION_BY_SCENARIO]},
     "dummy": {CONF_TRANSPORT: "dummy"},
 }